avoid crashing on bad input with qalc >= 5.0.0
Since version 5.0.0, qalc exits with code 1 when the input is invalid.
Reporting such a common problem (inevitable while typing) with g_error
is undesired, as calling g_error is fatal, and even if it weren't, this
would only lead to useless spamming of the standard error.

This commit also fixes related use-after-free in the same function.